+++ packages/react-reconciler/src/MaxInts.js
@@ -10,4 +10,4 @@
 // Max 31 bit integer. The max integer size in V8 for 32-bit systems.
 // Math.pow(2, 30) - 1
 // 0b111111111111111111111111111111
-export default 1073741823;
+export const MAX_SIGNED_31_BIT_INT = 1073741823;
