+++ packages/react/src/BadMapPolyfill.js
