+++ packages/react/src/React.js
@@ -27,9 +27,9 @@ import {
 } from './ReactElement';
 import {createContext} from './ReactContext';
 import {lazy} from './ReactLazy';
-import forwardRef from './forwardRef';
-import memo from './memo';
-import block from './block';
+import {forwardRef} from './ReactForwardRef';
+import {memo} from './ReactMemo';
+import {block} from './ReactBlock';
 import {
   useCallback,
   useContext,
@@ -52,11 +52,11 @@ import {
   createFactoryWithValidation,
   cloneElementWithValidation,
 } from './ReactElementValidator';
-import createMutableSource from './createMutableSource';
+import {createMutableSource} from './ReactMutableSource';
 import ReactSharedInternals from './ReactSharedInternals';
-import createFundamental from 'shared/createFundamentalComponent';
-import createResponder from 'shared/createEventResponder';
-import createScope from 'shared/createScope';
+import {createFundamental} from './ReactFundamental';
+import {createEventResponder} from './ReactEventResponder';
+import {createScope} from './ReactScope';
 
 // TODO: Move this branching into the other module instead and just re-export.
 const createElement = __DEV__ ? createElementWithValidation : createElementProd;
@@ -112,7 +112,7 @@ export {
   block,
   // enableDeprecatedFlareAPI
   useResponder as DEPRECATED_useResponder,
-  createResponder as DEPRECATED_createResponder,
+  createEventResponder as DEPRECATED_createResponder,
   // enableFundamentalAPI
   createFundamental as unstable_createFundamental,
   // enableScopeAPI
