+++ packages/react/src/ReactScope.js
@@ -9,7 +9,7 @@
 import type {ReactScope} from 'shared/ReactTypes';
 import {REACT_SCOPE_TYPE} from 'shared/ReactSymbols';
 
-export default function createScope(): ReactScope {
+export function createScope(): ReactScope {
   const scopeComponent = {
     $$typeof: REACT_SCOPE_TYPE,
   };
